Fixed the bug that led to different solutions in parallel and serial. It was the size of our solution buffers in the heat equation class. I need to add extra checks to make sure that vectors classified as "solution buffers" are length == nb_nodes, while the vectors classified as derivative buffers are length == nb_stencils

+++ src/utils/comm/mpisendable.cpp
@@ -522,7 +522,7 @@ int MPISendable::recvSTL(std::set<std::vector<Vec3> > *destination, int myrank, 
 	for (int i=0; i < sz; i++) {
 		std::vector<Vec3> temp; 
 		for (int j=0; j < buff[i]; j++) {
-			Vec3 v((double)buff2[offset+j][0], (double)buff2[offset+j][1], (double)buff2[offset+j][2]); // cast to ensure correct constructor
+			Vec3 v(buff2[offset+j][0], buff2[offset+j][1], buff2[offset+j][2]); // cast to ensure correct constructor
 			temp.push_back(v);
 		}
 		offset += buff[i];
